fix: changes in searchgroups and bugs in mygroups fixed
